requirements.txt
boto3~=1.21.21
botocore~=1.24.21
Flask~=2.1.1
PyJWT~=1.7.1
python-dotenv~=0.20.0